# specs/001-furniture-quotation-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 家具報價單系統 API
  description: |
    家具報價單系統後端 API，提供 PDF 上傳、BOQ 解析、Excel 報價單產出功能。

    ## 功能概述
    - PDF 檔案上傳與管理
    - BOQ 表格自動解析（使用 Gemini AI）
    - 平面圖數量核對
    - 惠而蒙格式 Excel 報價單產出

    ## 錯誤處理
    所有錯誤回應皆使用繁體中文訊息。
  version: 1.0.0
  contact:
    name: API 支援
servers:
  - url: http://localhost:8000
    description: 開發環境
  - url: http://api.example.com
    description: 生產環境
...
paths:
...
  /api/quotations:
    post:
      tags:
        - Export
      summary: 建立報價單
      description: 從解析結果建立報價單
      operationId: createQuotation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                document_ids:
                  type: array
                  items:
                    type: string
                  description: 來源文件 ID 列表
                title:
                  type: string
                  description: 報價單標題
              required:
                - document_ids
      responses:
        '201':
          description: 報價單已建立
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quotation'
  /api/quotations/{quotation_id}:
    get:
      tags:
        - Export
      summary: 取得報價單
      operationId: getQuotation
      parameters:
        - $ref: '#/components/parameters/QuotationId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Quotation'
        '404':
          description: 找不到報價單
  /api/quotations/{quotation_id}/items:
    get:
      tags:
        - Export
      summary: 取得報價單項目列表
      operationId: getQuotationItems
      parameters:
        - $ref: '#/components/parameters/QuotationId'
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/BOQItem'
                  total:
                    type: integer

    patch:
      tags:
        - Export
      summary: 更新報價單項目
      description: 批次更新報價單中的項目資料
      operationId: updateQuotationItems
      parameters:
        - $ref: '#/components/parameters/QuotationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                updates:
                  type: array
                  items:
                    type: object
                    properties:
                      id:
                        type: string
                      qty:
                        type: number
                      description:
                        type: string
                      materials_specs:
                        type: string
      responses:
        '200':
          description: 更新成功
  /api/quotations/{quotation_id}/excel:
    get:
      tags:
        - Export
      summary: 產出並下載 Excel 報價單
      description: |
        將報價單匯出為惠而蒙格式 Excel 檔案。
        若檔案尚未產出，將非同步產生並返回 202；
        若檔案已產出，直接返回 200 與檔案內容。
      operationId: exportExcel
      parameters:
        - $ref: '#/components/parameters/QuotationId'
        - name: include_photos
          in: query
          schema:
            type: boolean
            default: true
          description: 是否嵌入圖片
        - name: photo_height_cm
          in: query
          schema:
            type: number
            default: 3.0
          description: 圖片高度（公分）
      responses:
        '200':
          description: Excel 檔案
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '202':
          description: 匯出任務進行中，稍後重試
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskResponse'
        '404':
          description: 找不到報價單
...
components:
  parameters:
...
    QuotationId:
      name: quotation_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
      description: 報價單 ID
...
  schemas:
...
    BOQItem:
      type: object
      description: |
        BOQ 項目資料模型（完全比照惠而蒙格式 15 欄）
        Excel 欄位: A:NO., B:Item no., C:Description, D:Photo, E:Dimension,
        F:Qty, G:UOM, H:Unit Rate(留空), I:Amount(留空), J:Unit CBM,
        K:Total CBM(公式), L:Note, M:Location, N:Materials/Specs, O:Brand
      required:
        - id
        - item_no
        - description
      properties:
        id:
          type: string
          format: uuid
          description: 唯一識別碼（內部使用）
        no:
          type: integer
          minimum: 1
          description: A 欄：序號
        item_no:
          type: string
          description: B 欄：項目編號
        description:
          type: string
          description: C 欄：品名描述
        photo_base64:
          type: string
          description: D 欄：圖片 Base64 編碼
          nullable: true
        dimension:
          type: string
          description: E 欄：尺寸 WxDxH (mm)
          nullable: true
        qty:
          type: number
          description: F 欄：數量
          nullable: true
        uom:
          type: string
          description: G 欄：單位（ea, m, set 等）
          nullable: true
        unit_cbm:
          type: number
          description: J 欄：單位材積
          nullable: true
        note:
          type: string
          description: L 欄：備註
          nullable: true
        location:
          type: string
          description: M 欄：位置
          nullable: true
        materials_specs:
          type: string
          description: N 欄：使用材料/規格
          nullable: true
        brand:
          type: string
          description: O 欄：品牌
          nullable: true
        source_document_id:
          type: string
          description: 來源文件 ID（內部使用）
        source_page:
          type: integer
          description: 來源頁碼（內部使用）
          nullable: true
...
    Quotation:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        created_at:
          type: string
          format: date-time
        source_document_ids:
          type: array
          items:
            type: string
        total_items:
          type: integer
        items_with_qty:
          type: integer
        items_with_photo:
          type: integer
        items_from_floor_plan:
          type: integer
        export_status:
          type: string
          enum: [pending, generating, completed, failed]
        export_path:
          type: string
          nullable: true
    TaskResponse:
      type: object
      properties:
        task_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [pending, processing, completed, failed]
        message:
          type: string
          example: 任務已建立，正在處理中
...